+++ pkg/watcher/file_watcher_test.go
